"refactor: split controller (NewInsuranceRequest), reorganize file structure"
