+++ src/main/java/ru/ya/insurance/mapper/newInsuranceRequest/NewInsuranceRequestMapper.java
@@ -1,8 +1,8 @@
-package ru.ya.insurance.mapper.osago;
+package ru.ya.insurance.mapper.newInsuranceRequest;
 
 import org.mapstruct.Mapper;
-import ru.ya.insurance.dto.osago.NewInsuranceRequestDto;
-import ru.ya.insurance.model.osago.NewInsuranceRequest;
+import ru.ya.insurance.dto.newInsuranceRequest.NewInsuranceRequestDto;
+import ru.ya.insurance.model.newInsuranceRequest.NewInsuranceRequest;
 
 @Mapper
 public interface NewInsuranceRequestMapper {
