+++ src/main/java/ru/ya/insurance/repository/newInsuranceRequest/NewInsuranceRequestRepository.java
@@ -1,8 +1,8 @@
-package ru.ya.insurance.repository.osago;
+package ru.ya.insurance.repository.newInsuranceRequest;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.ya.insurance.model.osago.NewInsuranceRequest;
+import ru.ya.insurance.model.newInsuranceRequest.NewInsuranceRequest;
 
 @Repository
 public interface NewInsuranceRequestRepository extends JpaRepository<NewInsuranceRequest, Long> {
