+++ src/main/java/ru/ya/insurance/service/osago/impl/NewInsuranceRequestServiceImpl.java
@@ -3,11 +3,11 @@ package ru.ya.insurance.service.osago.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.ya.insurance.dto.osago.NewInsuranceRequestDto;
-import ru.ya.insurance.mapper.osago.NewInsuranceRequestMapper;
-import ru.ya.insurance.model.osago.NewInsuranceRequest;
-import ru.ya.insurance.repository.osago.NewInsuranceRequestRepository;
-import ru.ya.insurance.service.osago.NewInsuranceRequestService;
+import ru.ya.insurance.dto.newInsuranceRequest.NewInsuranceRequestDto;
+import ru.ya.insurance.mapper.newInsuranceRequest.NewInsuranceRequestMapper;
+import ru.ya.insurance.model.newInsuranceRequest.NewInsuranceRequest;
+import ru.ya.insurance.repository.newInsuranceRequest.NewInsuranceRequestRepository;
+import ru.ya.insurance.service.newInsuranceRequest.NewInsuranceRequestService;
 
 @Transactional(readOnly = true)
 @Service
